Move stubs to corresponding class in repository test-suite instead of separate files.
